Fixed bug in size calculation.
git-svn-id: https://fractalfrost.svn.sourceforge.net/svnroot/fractalfrost/trunk@429 24c11add-7f9c-4b28-9183-29c26a6ab64e

+++ lib/gui/renderdialog.py
@@ -280,7 +280,7 @@ class RenderDialog(wx.Frame):
             else:
                 wtc.SetInt(v * self.ratio)
         else:
-            self.ratio = wtc.GetFloat() / htc.GetFloat()
+            self.ratio = float(wtc.GetInt()) / htc.GetInt()
         self.mem.UpdateView()
